+++ Data_Vanilla.lua
@@ -1,5 +1,5 @@
 local _, _addon = ...
-local buildNr = 45437
+local buildNr = 46575
 
 -- \/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/