+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 10.0.02
+## 10.0.03
 
-- Updated Retail atlas data for patch 10.0.2 (build 46479).
+- Updated Retail atlas data for build 10.0.5.48317.
+- Updated Classic atlas data for build 3.4.1.48340.
+- Updated Classic Era atlas data for build 1.14.3.46575.